feat(dReach): add option to filter encodings
The current approach is to always include the encodings of the whole network.
We introduce an options to change the behavior to filter out encodings of modes and related contraints, that are unreachable within k steps.
Execute either the dReach runscript with the -f option, or the bmc executable with the --length_filter options.

fix(style)